requirements.txt
requests
openai
python-dotenv
flake8
pyyaml